Started testing for PrepareDownload
