+++ presqt/api_v1/tests/views/resource/test_resource_collection.py
@@ -1,0 +1,82 @@
+from unittest.mock import patch
+
+from django.test import TestCase
+from rest_framework.reverse import reverse
+from rest_framework.test import APIClient
+
+from config.settings.base import TEST_USER_TOKEN
+
+class TestResourceCollection(TestCase):
+    """
+    Test the 'api_v1/targets/{target_name}/resources/' endpoint's GET method.
+    """
+
+    def setUp(self):
+        self.client = APIClient()
+        self.header = {'HTTP_PRESQT_SOURCE_TOKEN': TEST_USER_TOKEN}
+
+    def test_get_success_osf(self):
+        """
+        Return a 200 if the GET method is successful when grabbing OSF resources.
+        """
+        url = reverse('resource_collection', kwargs={'target_name': 'osf'})
+        response = self.client.get(url, **self.header)
+        # Verify the Status Code
+        self.assertEqual(response.status_code, 200)
+        # Verify the dict keys match what we expect
+        keys = ['kind', 'kind_name', 'id', 'container', 'title']
+        for data in response.data:
+            self.assertListEqual(keys, list(data.keys()))
+        # Verify the count of resource objects is what we expect.
+        self.assertEqual(69, len(response.data))
+
+    def test_get_error_400_missing_token_osf(self):
+        """
+        Return a 400 if the GET method fails because the presqt-source-token was not provided.
+        """
+        url = reverse('resource_collection', kwargs={'target_name': 'osf'})
+        response = self.client.get(url)
+        # Verify the error status code and message
+        self.assertEqual(response.status_code, 400)
+        self.assertEqual(response.data,
+                         {'error': "'presqt-source-token' missing in the request headers."})
+
+    def test_get_error_400_target_not_supported_test_target(self):
+        """
+        Return a 400 if the GET method fails because the target requested does not support
+        this endpoint's action.
+        """
+        with open('presqt/api_v1/tests/views/targets_test.json') as json_file:
+            with patch("builtins.open") as mock_file:
+                mock_file.return_value = json_file
+                url = reverse('resource_collection', kwargs={'target_name': 'test'})
+                response = self.client.get(url, **self.header)
+                # Verify the error status code and message
+                self.assertEqual(response.status_code, 400)
+                self.assertEqual(
+                    response.data,
+                    {'error': "'test' does not support the action 'resource_collection'."})
+
+
+    def test_get_error_401_invalid_token_osf(self):
+        """
+`       Return a 401 if the token provided is not a valid token.
+        """
+        client = APIClient()
+        header = {'HTTP_PRESQT_SOURCE_TOKEN': 'bad_token'}
+        url = reverse('resource_collection', kwargs={'target_name': 'osf'})
+        response = client.get(url, **header)
+        # Verify the error status code and message
+        self.assertEqual(response.status_code, 401)
+        self.assertEqual(response.data,
+                         {'error': "Token is invalid. Response returned a 401 status code."})
+
+    def test_get_error_404_bad_target_name_osf(self):
+        """
+        Return a 404 if the GET method fails because a bad target_name was given.
+        """
+        url = reverse('resource_collection', kwargs={'target_name': 'bad_name'})
+        response = self.client.get(url, **self.header)
+        # Verify the error status code and message
+        self.assertEqual(response.status_code, 404)
+        self.assertEqual(response.data, {'error': "'bad_name' is not a valid Target name."})
