+++ presqt/api_v1/urls.py
@@ -1,8 +1,10 @@
-from django.urls import path, re_path
+from django.urls import path
 
 from presqt.api_v1 import api_root
-from presqt.api_v1.views.resource import ResourceCollection, Resource, PrepareDownload
-from presqt.api_v1.views.target import TargetCollection, Target
+from presqt.api_v1.views.resource.resource import Resource
+from presqt.api_v1.views.resource.resource_collection import ResourceCollection
+from presqt.api_v1.views.resource.resource_download import PrepareDownload
+from presqt.api_v1.views.target.target import TargetCollection, Target
 
 api_v1_endpoints = [
     path('', api_root, name='api_root'),
